Implement decimation
This commit implement the decimation algorithm from the paper "Interactively Controlled Quad Remeshing of High Resolution 3D Models".

The coarse to fine mapping algorithm described in the paper is been implemented in decimated branch, but the parametrization UV after coarse to fine mapping is too distorted to be useful, the mesh generated from that UV is distored as well. This remains to be further investigated.

Luckily, as been observed, the target decimated vertex count of 5000 is already produced good mesh without coarse to fine mapping. Instead of decimating to 1000 vertices and mapping back to a fine level, we can just decimate to 5000 vertices without mapping back.